Initial Commit
Includes:

- command.js
Allows for use of command method which takes in a client, aliases, and a callback.

- index.js
Main file. Makes use of the command function from commad.js

- Run.bat
Executable file to start the bot

- config.json
Not included. Just has a prefix and token. The token is exclusive to every user's bot. The prefix can be changed to anything, but in this case, it's "$".